ref(preprod): Extract artifact search functions to artifact_search module
